requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
argon2-cffi
python-jose
email-validator
pydantic
langdetect
langcodes
googletrans==4.0.0-rc1
requests
asyncio
joblib
scikit-learn
xgboost
catboost
matplotlib
seaborn
python-dotenv